% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils_fns.R
\name{auto_log}
\alias{auto_log}
\title{auto log2 transform for GEO datasets}
\usage{
auto_log(data)
}
\arguments{
\item{data}{gene expression data, can be a \code{ExpressionSet} object  or
\code{matrix} object.}
}
\value{
a \code{matrix} or \code{ExpressionSet} object of logarithm
transformed if it is unlogged
\code{class(data_set)}
}
\description{
auto log2 transform for GEO datasets
}
\details{
Automatically check whether \code{data} has experientced logarithm
transformation, if not, applying a log2 transformation. The test
methodology for logarithm transformation is based on
\href{https://www.ncbi.nlm.nih.gov/geo/geo2r/}{GEO2R}
}
\references{
\href{https://www.ncbi.nlm.nih.gov/geo/geo2r/?acc=GSE1122}{GEO
analysis}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
