% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_c}
\alias{cols_c}
\title{select columns using tidyselect}
\usage{
cols_c(..., data)
}
\arguments{
\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more
unquoted expressions separated by commas. Variable names can be used as if
they were positions in the data frame, so expressions like \code{x:y} can be
used to select a range of variables.}

\item{data}{A data frame or data.table}
}
\description{
select columns using tidyselect
}
