% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_num}
\alias{format_num}
\title{format a numeric vector}
\usage{
format_num(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
A character vector containing the formatted x
}
\description{
format a numeric vector
}
