% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstat_utils_fns.R
\name{gganno_between_test}
\alias{gganno_between_test}
\title{Create a ggplot2 object with annotation for group or condition test results}
\usage{
gganno_between_test(
  data,
  mapping = NULL,
  type = c("nonparametric", "parametric"),
  format_label = TRUE,
  label_position = c(0.02, 0.98),
  label_justification = c(0, 1),
  anno_statistic = TRUE,
  label_sep = NULL,
  anno_args = list(fontface = "bold"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not specified, must be
supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. x and y
aesthetic must in.}

\item{type}{if \code{type = "nonparametric"},
\code{\link[stats]{wilcox.test}} or \code{\link[stats]{kruskal.test}} is
used. \cr if \code{type = "parametric"}, \code{\link[stats]{t.test}} or
\code{\link[stats]{aov}} is used.\cr which method is used depends on the
unique values of \code{x} in \code{data}}

\item{format_label}{if \code{TRUE}, numeric label will be formatted by
\code{\link{format_num}}. Default: \code{TRUE}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length two or a character vector with length one (one of
\code{"lefttop", "righttop", "leftbottom", and "rightbottom"}) or length
two (See below \code{label_justification}). Default: \code{c(0.02, 0.98)}.}

\item{label_justification}{label_justification must be a element-two (or one)
character or numeric vector. Possible string values are: "left", "right",
"centre", "center", "bottom", and "top". For numeric values, 0 means left
(bottom) alignment and 1 means right (top) alignment. Default: \code{c(0,
  1)}}

\item{anno_statistic}{a scalar logical, Should test statistic be annotated in
the label}

\item{label_sep}{String to insert between the test statistic and test P-value.}

\item{anno_args}{Other arguments passed on to \code{\link[ggplot2]{layer}}.
These are often aesthetics, used to set an aesthetic to a fixed value, like
colour = "red" or size = 3. They may also be parameters to the paired
geom/stat. See \code{\link[ggplot2:geom_text]{geom_text}}}

\item{...}{other arguments passed to statistical test function,
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{kruskal.test}},
\code{\link[stats]{t.test}} and \code{\link[stats]{aov}}.}
}
\value{
a ggplot2 object with annotation of group or condition comparisons
test results
}
\description{
annotate group or condition comparisons in between-subjects designs in a
ggplot2 object.
}
\examples{
gganno_between_test(mtcars, aes(factor(cyl), mpg)) +
  ggplot2::geom_boxplot()
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
