% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils_fns.R
\name{gganno_text}
\alias{gganno_text}
\title{Create an ggplot2 object with annotation}
\usage{
gganno_text(
  data,
  mapping = NULL,
  label,
  format_label = TRUE,
  label_position = c(0.02, 1),
  label_justification = NULL,
  label_sep = NULL,
  ggtheme = ggthemes::theme_few(),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not specified, must be
supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each layer added to the plot.}

\item{label}{the character which will be added to the plot, will be connected
by \code{stringr::str_c(names(label), label), collapse = label_sep, sep =
  ": "}}

\item{format_label}{if \code{TRUE}, will be formatted by
\code{\link{format_num}}. Default: \code{TRUE}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length 2 or a character vector \code{(one of "title",
  "subtitle", "caption", and "tag")} of length one.}

\item{label_justification}{label_justification must be a element-two (or one)
character or numeric vector. Possible string values are: "left", "right",
"centre", "center", "bottom", and "top". For numeric values, 0 means left
(bottom) alignment and 1 means right (top) alignment. Details see
\code{\link[grid]{valid.just}}. Default: for numeric \code{label_position},
default \code{label_justification} is \code{c(0, 1.15)}, for character
\code{label_position}, default \code{label_justification} is \code{c(0, 0)}}

\item{label_sep}{see argument \code{label}. String to insert between each of
the \code{label}.}

\item{ggtheme}{a ggplot2 theme object. Provided as a ggplot2 theme object or
a function (can be a function name as a string) implemented as a ggplot2
theme object.}

\item{...}{additional arguments control the \code{label}. See
\code{\link[grid]{textGrob}} and \code{\link[ggplot2]{element_text}}}
}
\value{
a ggplot2 object with annotation
}
\description{
This function Creates an ggplot2 object with annotation.
}
\examples{
  gganno(mtcars, aes(cyl, mpg), label = "a")
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
