% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils_fns.R
\name{ggannotate_npc}
\alias{ggannotate_npc}
\title{Create an annotation layer}
\usage{
ggannotate_npc(geom, x, y, na.rm = FALSE, ...)
}
\arguments{
\item{geom}{character Name of geom to use for annotation. Can be one of
\code{"grob"}, \code{"plot"}, \code{"text"}, \code{"label"}, and
\code{"table"}. Details see \code{\link[ggpmisc]{annotate}}}

\item{x, y}{the position of added label in Normalised Parent
Coordinates(\code{npc})}
}
\description{
This function adds geoms to a plot with x and y mapping to Normalised Parent
Coordinates
}
\details{
See \code{\link{ggplot2::annotate}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
