% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate_npc.R
\name{ggannotate_npc_helper}
\alias{ggannotate_npc_helper}
\title{Create an ggplot2 annotation layer}
\usage{
ggannotate_npc_helper(geom, x = NULL, y = NULL, ..., na.rm = FALSE)
}
\arguments{
\item{geom}{character Name of geom to use for annotation. Can be one of
\code{"text"} and \code{"label"}.}

\item{x, y}{the position of added label in Normalised Parent
Coordinates(\code{npc})}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}}. These
are often aesthetics, used to set an aesthetic to a fixed value, like
colour = "red" or size = 3. They may also be parameters to the paired
geom/stat. See \code{\link[ggplot2:geom_text]{geom_text}} and
\code{\link[ggplot2:geom_text]{geom_label}}.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}
}
\description{
This function adds geoms to a plot with x and y mapping to Normalised Parent
Coordinates
}
\details{
See \code{\link[ggplot2]{annotate}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
