% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_extensions_label_npc.R,
%   R/ggplot2_extensions_text_npc.R
\docType{data}
\name{GeomLabelNpc}
\alias{GeomLabelNpc}
\alias{GeomTextNpc}
\title{Base ggproto classes}
\description{
This geom is identical to 'ggplot2' \code{\link[ggplot2]{geom_text}}
and  except that it interprets \code{x} and \code{y} positions in
\code{npc} units. It translates \code{x} and \code{y} coordinates from npc
units to native data units.
}
\details{
See \code{\link[ggplot2]{ggplot2-ggproto}}
}
\seealso{
\code{\link[ggplot2]{ggproto}}
}
\keyword{datasets}
