% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils_fns.R
\name{ggscale_paletteer}
\alias{ggscale_paletteer}
\title{Sequential, diverging and qualitative colour scales from \code{paletteer}}
\usage{
ggscale_paletteer(
  palette = "nejm",
  scale = "colour",
  type = c("discrete", "continuous", "binned"),
  direction = 1,
  ...
)
}
\arguments{
\item{palette}{Name of palette as a string. Can be in the form of
\code{packagename::palettename}. Details see
\code{\href{https://emilhvitfeldt.github.io/paletteer/}{paletteer}}. for
\code{palette} in \code{ggsci}, we will use regular expression to match. If
match nothing in \code{paletteer palette}.
\code{\link[ggplot2:scale_manual]{scale_discrete_manual}} will be used.}

\item{scale}{the aesthetic mapping the palette}

\item{type}{the type of scale, a scalar character, can be "\code{discrete}",
"\code{continuous}", and "\code{binned}". Default: "\code{discrete}"}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as output by
\code{\link[paletteer:ggplot2-scales-continuous]{paletteer}}. If -1, the
order of colours is reversed. Default: \code{1}.}

\item{...}{Additional arguments pass on to
\code{\link[ggplot2:discrete_scale]{discrete scale}},
\code{\link[ggplot2:scale_gradient]{gradient continuous scale}} or
\code{\link[ggplot2:scale_steps]{stepsn scale}}.}
}
\value{
A Scale object that can be added to a ggplot object
}
\description{
The choices of color palettes in R can be quite overwhelming
with palettes spread over many packages with many different API's. This
packages aims to collect all color palettes across the R ecosystem under
the same package with a streamlined API. See
\code{\link[paletteer]{paletteer-package}}
}
\examples{
  ggscale_paletteer()
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
