% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstat_utils_fns.R
\name{ggstat_anno_helper}
\alias{ggstat_anno_helper}
\title{Create an annotation layer with text annotation}
\usage{
ggstat_anno_helper(
  geom,
  label,
  format_label = TRUE,
  label_position = c(0.02, 0.98),
  label_justification = c(0, 1),
  label_sep = NULL,
  ...
)
}
\arguments{
\item{geom}{character Name of geom to use for annotation. Can be one of
\code{"text"} and \code{"label"}.}

\item{label}{the character which will be added to the plot, will be connected
by \code{paste(names(label), label, collapse = label_sep, sep = ": ")}}

\item{format_label}{if \code{TRUE}, numeric label will be formatted by
\code{\link{format_num}}. Default: \code{TRUE}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length two or a character vector with length one (one of
\code{"lefttop", "righttop", "leftbottom", and "rightbottom"}) or length
two (See below \code{label_justification}). Default: \code{c(0.02, 0.98)}.}

\item{label_justification}{label_justification must be a element-two (or one)
character or numeric vector. Possible string values are: "left", "right",
"centre", "center", "bottom", and "top". For numeric values, 0 means left
(bottom) alignment and 1 means right (top) alignment. Default: \code{c(0,
  1)}}

\item{label_sep}{see argument \code{label}. String to insert between each of
the \code{label}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}}.
These are often aesthetics, used to set an aesthetic to a fixed value, like
colour = "red" or size = 3. They may also be parameters to the paired
geom/stat. See \code{\link[ggplot2:geom_text]{geom_text}}}
}
\value{
a ggplot2 annotation layer
}
\description{
This function Creates an ggplot2 object with annotation.
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
