% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Statistical Inference with Permutation}
\usage{
infer(
  data,
  formula,
  stat,
  reps = 2000L,
  level = 0.95,
  direction = "two-sided",
  null = NULL,
  type = NULL,
  null_reps = reps,
  ci_reps = reps,
  ...,
  variables = NULL,
  response = NULL,
  explanatory = NULL,
  success = NULL,
  p = NULL,
  mu = NULL,
  med = NULL,
  sigma = NULL
)
}
\arguments{
\item{data}{A data frame that can be coerced into a tibble.}

\item{formula}{A formula with the response variable on the left and the
explanatory on the right. Alternatively, a \code{response} and \code{explanatory}
argument can be supplied.}

\item{stat}{A string giving the type of the statistic to calculate. Current
options include \code{"mean"}, \code{"median"}, \code{"sum"}, \code{"sd"}, \code{"prop"}, \code{"count"},
\code{"diff in means"}, \code{"diff in medians"}, \code{"diff in props"}, \code{"Chisq"} (or
\code{"chisq"}), \code{"F"} (or \code{"f"}), \code{"t"}, \code{"z"}, \code{"ratio of props"}, \code{"slope"},
\code{"odds ratio"}, \code{"ratio of means"}, and \code{"correlation"}. \code{infer} only
supports theoretical tests on one or two means via the \code{"t"} distribution
and one or two proportions via the \code{"z"}.}

\item{reps}{Fast way to set both \code{null_reps} and \code{ci_reps}.}

\item{level}{A numerical value between 0 and 1 giving the confidence level.
Default value is 0.95.}

\item{direction}{A character string. Options are \code{"less"}, \code{"greater"}, or
\code{"two-sided"}. Can also use \code{"left"}, \code{"right"}, \code{"both"},
\code{"two_sided"}, or \code{"two sided"}, \code{"two.sided"}.}

\item{null}{The null hypothesis. Options include \code{"independence"},
\code{"point"}, and \code{"paired independence"}.
\itemize{
\item \code{independence}: Should be used with both a \code{response} and \code{explanatory}
variable. Indicates that the values of the specified \code{response} variable
are independent of the associated values in \code{explanatory}.
\item \code{point}: Should be used with only a \code{response} variable. Indicates
that a point estimate based on the values in \code{response} is associated
with a parameter. Sometimes requires supplying one of \code{p}, \code{mu}, \code{med}, or
\code{sigma}.
\item \verb{paired independence}: Should be used with only a \code{response} variable
giving the pre-computed difference between paired observations. Indicates
that the order of subtraction between paired values does not affect the
resulting distribution.
}}

\item{type}{A string giving which method should be used for creating the
confidence interval. The default is \code{"percentile"} with \code{"se"}
corresponding to (multiplier * standard error) and \code{"bias-corrected"} for
bias-corrected interval as other options.}

\item{null_reps}{Number of times to calculate null distribution.}

\item{ci_reps}{Number of bootstrap times to calculate confidence interval.}

\item{...}{
  Arguments passed on to \code{\link[infer:calculate]{infer::calculate}}
  \describe{
    \item{\code{order}}{A string vector of specifying the order in which the levels of
the explanatory variable should be ordered for subtraction (or division
for ratio-based statistics), where \code{order = c("first", "second")} means
\code{("first" - "second")}, or the analogue for ratios. Needed for inference on
difference in means, medians, proportions, ratios, t, and z statistics.}
  }}

\item{variables}{A set of unquoted column names in the data to permute
(independently of each other). Defaults to only the response variable. Note
that any derived effects that depend on these columns (e.g., interaction
effects) will also be affected.}

\item{response}{The variable name in \code{x} that will serve as the response.
This is an alternative to using the \code{formula} argument.}

\item{explanatory}{The variable name in \code{x} that will serve as the
explanatory variable. This is an alternative to using the formula argument.}

\item{success}{The level of \code{response} that will be considered a success, as
a string. Needed for inference on one proportion, a difference in
proportions, and corresponding z stats.}

\item{p}{The true proportion of successes (a number between 0 and 1). To be used with point null hypotheses when the specified response
variable is categorical.}

\item{mu}{The true mean (any numerical value). To be used with point null
hypotheses when the specified response variable is continuous.}

\item{med}{The true median (any numerical value). To be used with point null
hypotheses when the specified response variable is continuous.}

\item{sigma}{The true standard deviation (any numerical value). To be used with
point null hypotheses.}
}
\value{
A data.frame
}
\description{
Statistical Inference with Permutation
}
\examples{
data(gss, package = "infer")
infer(gss, response = hours, stat = "mean", mu = 40)
infer(gss, response = hours, stat = "t", mu = 40)
infer(gss, response = hours, stat = "median", med = 40)
infer(gss, response = sex, success = "female", stat = "prop", p = .5)
infer(gss, response = sex, success = "female", stat = "z", p = .5)
infer(gss, college ~ sex,
  success = "no degree",
  stat = "diff in props",
  order = c("female", "male")
)
infer(gss, hours ~ age + college, variables = c(age, college))
}
\seealso{
\url{https://infer.tidymodels.org/articles/observed_stat_examples.html}
}
