% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Install or update Bioconductor, CRAN and GitHub packages}
\usage{
install_pkgs(
  pkgs = character(),
  ...,
  type = "source",
  INSTALL_opts = "--byte-compile",
  site_repository = character(),
  update = TRUE,
  ask = TRUE,
  checkBuilt = FALSE
)
}
\arguments{
\item{pkgs}{\code{character()} vector of package names to install or
update.  A missing value updates installed packages according
to \verb{update =} and \verb{ask =}. Package names containing a '/' are
treated as GitHub repositories and installed using
\code{remotes::install_github()}.}

\item{...}{Additional arguments used by \code{install.packages()}.}

\item{site_repository}{(Optional) \code{character(1)} vector
representing an additional repository in which to look for
packages to install. This repository will be prepended to the
default repositories (which you can see with
\verb{BiocManager::\\link{repositories}()}).}

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{ask}{\code{logical(1)} indicating whether to prompt user before
installed packages are updated.  If TRUE, user can choose
whether to update all outdated packages without further
prompting, to pick packages to update, or to cancel updating
(in a non-interactive session, no packages will be updated
unless \code{ask = FALSE}).}

\item{checkBuilt}{\code{logical(1)}. If \code{TRUE} a package built under an
earlier major.minor version of R (e.g., 3.4) is considered to
be old.}
}
\description{
Install or update CRAN, Bioconductor and GitHub packages from source and
compiling
}
