% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_absolute.R
\name{run_absolute}
\alias{run_absolute}
\title{Automate ABSOLUTE calling for multiple samples in parallel ways}
\usage{
run_absolute(
  seg,
  maf = NULL,
  sigma_p = 0,
  max_sigma_h = 0.015,
  min_ploidy = 0.95,
  max_ploidy = 10,
  primary_disease = NA,
  platform = c("SNP_6.0", "Illumina_WES", "SNP_250K_STY"),
  temp_dir = file.path(tempdir(), "ABSOLUTE"),
  clean_temp = FALSE,
  results_dir = here::here("results", "ABSOLUTE"),
  max_as_seg_count = 1500,
  max_neg_genome = 0.005,
  max_non_clonal = 0.05,
  copy_num_type = c("total", "allelic"),
  min_mut_af = 0.1,
  keep_all_results = TRUE,
  recover = FALSE,
  verbose = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{seg}{a \code{data.frame} or a \code{file path} read as a data.fram
containing columns "Sample", "Chromosome", "Start", "End", "Num_Probes",
"Segment_Mean".}

\item{maf}{MAF, default is \code{NULL}, can provided as \code{data.frame} or
\code{file path}.}

\item{sigma_p}{Provisional value of excess sample level variance used for
mode search. Default: \code{0}}

\item{max_sigma_h}{Maximum value of excess sample level variance (Eq. 6).
Default: \code{0.015}}

\item{min_ploidy}{Minimum ploidy value to consider. Solutions implying lower
ploidy values will be discarded. Default: \code{0.95}}

\item{max_ploidy}{Maximum ploidy value to consider. Solutions implying
greater ploidy values will be discarded. Default: \code{10}}

\item{primary_disease}{Primary disease of the sample. Default: \code{NA}}

\item{platform}{one of "SNP_6.0", "Illumina_WES", "SNP_250K_STY". Default:
\code{"SNP_6.0"}}

\item{temp_dir}{directory path used to store tempory files. Default: ABSOLUTE
subdirectory under \code{tempdir()}}

\item{clean_temp}{if \code{TRUE}, auto-clean temp dir at the end. Default:
\code{FALSE}}

\item{results_dir}{directory path used to store result files. Default:
\code{here::here("results", "ABSOLUTE")}}

\item{max_as_seg_count}{Maximum number of allelic segments. Samples with a
higher segment count will be flagged as 'failed'. Default: \code{1500}}

\item{max_neg_genome}{Maximum genome fraction that may be modeled as
non-clonal with copy-ratio below that of clonal homozygous deletion.
Solutions implying greater values will be discarded. Default: \code{0.005}}

\item{max_non_clonal}{Maximum genome fraction that may be modeled as
non-clonal (subclonal SCNA). Solutions implying greater values will be
discarded. Default: \code{0.05}}

\item{copy_num_type}{The type of copy number to be handled. Either total or
allelic. total is what this code for. Default: \code{"total"}}

\item{min_mut_af}{Minimum mutation allelic fraction. Mutations with lower
allelic fractions will be filtered out before analysis. Default: \code{0.1}}

\item{keep_all_results}{if \code{TRUE}, clean all results, otherwise clean
result directory and keep most important results. Default: \code{TRUE}}

\item{recover}{if \code{TRUE}, recover previous unfinished work. This is
helpful when program stop unexpectedly when \code{clean.temp} is
\code{FALSE}. Default: \code{FALSE}}

\item{verbose}{if \code{TRUE}, print extra info. Default: \code{FALSE}}

\item{BPPARAM}{see \code{\link[BiocParallel:register]{bpparam}}. Default:
\code{NULL} means \cr \code{switch (Sys.info()[["sysname"]],} \cr
\code{Linux = BiocParallel::MulticoreParam(),} \cr \code{Windows =
  BiocParallel::SnowParam()) }}
}
\value{
All reviewed results (see
\code{\link[ABSOLUTE]{ExtractReviewedResults}}) were in directory
\code{results_dir}.

if keep_all_results was TRUE, all ABSOLUTE called results (see
\code{\link[ABSOLUTE]{RunAbsolute}}) were kept in directory
\code{file.path(results_dir, "call_res")}, all summarized multiple ABSOLUTE
results (see \code{\link[ABSOLUTE]{CreateReviewObject}}) were kept in
\code{file.path(results_dir, "summary_res")}, and all reviewed samples were
kept in \code{ file.path(results_dir, "sample_summaized")}
}
\description{
This function is based on package
\href{https://github.com/ShixiangWang/DoAbsolute}{DoAbsolute} by adding
parallel evaluation and adjusting personal convention
}
\details{
\href{https://www.nature.com/articles/nbt.2203}{ABSOLUTE} is a
famous software developed by Broad Institute, However, the
\code{\link[ABSOLUTE]{RunAbsolute}} function points to estimate one sample
each time and sets no default values. \code{\link{run_absolute}} helps
users set default parameters based on
\href{https://www.genepattern.org/modules/docs/ABSOLUTE}{ABSOLUTE
documentation} and provides an uniform interface to input data easily and
run \code{\link[ABSOLUTE]{RunAbsolute}} parallelly.

If calling for a sample failed, the error message will be written to
\strong{error.log} under the \code{results_dir} directory.

More detail about how to analyze ABSOLUTE results please see
\href{https://www.genepattern.org/analyzing-absolute-data}{analyzing-absolute-data}.
}
\examples{
\donttest{
seg <- readRDS(
system.file("extdata", "run_absolute_example_seg.rds", package = "yjkit")
)

maf <- readRDS(
system.file("extdata", "run_absolute_example_maf.rds", package = "yjkit")
)

run_absolute(seg = seg, maf = maf,
results_dir = file.path(tempdir(),"results", "ABSOLUTE")
)
}
}
\references{
Carter, S., Cibulskis, K., Helman, E. et al. Absolute
quantification of somatic DNA alterations in human cancer. Nat Biotechnol
30, 413–421 (2012). \url{https://doi.org/10.1038/nbt.2203}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
