% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arm_cnv.R
\name{run_arm_cnv}
\alias{run_arm_cnv}
\title{Caculate Chromosome-arm-levels copy number variation}
\usage{
run_arm_cnv(
  seg_cnv,
  cnv_col,
  sample_id_col = NULL,
  ref_cytoband = NULL,
  filter_centromere = TRUE,
  cnv_mode = c("rel", "abs"),
  ploidy = NULL,
  threshold = 0.9
)
}
\arguments{
\item{seg_cnv}{is a GenomicRanges obeject with interger values of CNV column
in \code{\link[S4Vectors]{mcols}(seg_cnv)[[cnv_col]]} (absolute segment
copy number determined by ABSOLUTE algorithm or relative segment copy
number defined with -1 meaning del, 0 meaning neutral and 1 meaning amp)
and samples ID in \code{\link[S4Vectors]{mcols}(seg_cnv)[[sample_id_col]]}}

\item{cnv_col}{a scalar character gives the column containing CNV}

\item{sample_id_col}{a scalar character gives the column containing sample ID}

\item{ref_cytoband}{is a GenomicRanges obeject containing the Cytoband
reference, It can be a scalar character "hg19" or "hg38", or you can
provided a self-defined GenomicRanges obeject. Default: "hg38"}

\item{filter_centromere}{Whether to include or exclude segments across
centromere. Default: \code{TRUE}}

\item{cnv_mode}{is a scala character with values in "rel" and "abs"
correspongding to what Shukla, A and Cohen-Sharir have presented
respectively}

\item{ploidy}{the ploidy for the comparison to define Chromosome arm-level
aneuploidy profiling. Cohen-Sharir uses background ploidy (\code{cnv_mode
  \strong{abs}})}

\item{threshold}{the fraction to define Chromosome arm-level aneuploidy
profiling,  Cohen-Sharir uses 0.9 as the cut-off. (\code{cnv_mode
  \strong{rel}})}
}
\value{
a tibble containing Chromosome-arm-levels (CNV)
}
\description{
Caculate Chromosome-arm-levels copy number variation
}
\examples{
seg_cnv <- readRDS(system.file("extdata", "run_arm_cnv_example_seg_cnv.rds",
package = "yjkit"))
arm_cnv_res <- run_arm_cnv(seg_cnv, "CNV", "barcode")
}
\references{
\itemize{\item   Cohen-Sharir, Y., McFarland, J.M., Abdusamad,
M. et al. Aneuploidy renders cancer cells vulnerable to mitotic checkpoint
inhibition. Nature 590, 486–491 (2021).
\url{https://doi.org/10.1038/s41586-020-03114-6} \item Shukla, A., Nguyen,
T.H.M., Moka, S.B. et al. Chromosome arm aneuploidies shape tumour
evolution and drug response. Nat Commun 11, 449 (2020).
\url{https://doi.org/10.1038/s41467-020-14286-0}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
