% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ssgsea.R
\name{run_ssgsea}
\alias{run_ssgsea}
\title{Single-sample GSEA}
\usage{
run_ssgsea(
  data_set,
  gene_set_list,
  assay,
  NES = TRUE,
  num_perm = 200,
  gene_set_selection = "ALL",
  min_sz = 1,
  max_sz = Inf,
  sample_norm_type = c("rank", "log", "log.rank", "none"),
  weight = 0.75,
  min_overlap = 10,
  BPPARAM = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_set}{gene expression data, can be \code{ExpressionSet} object,
\code{SummarizedExperiment}, or \code{matrix} object. when data_set is a
\code{SummarizedExperiment} object, the \code{assay} argument will be used
to extracted the assay with exression matrix (see
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}),
if missing \code{assay}, the first \code{assay} will be used. For ssGSEA,
normalizd exression values with gene length adjusted were needed.}

\item{gene_set_list}{gene sets can be provided as \code{GeneSetCollection}
object or \code{list} object}

\item{assay}{used to extracted the assay with exression matrix (see
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}),
if missing \code{assay}, the first \code{assay} will be used.}

\item{NES}{whether to calculate normalized enrichment scores}

\item{num_perm}{the number of permutations to calculate NES}

\item{gene_set_selection}{list of gene set names on which to project the
input expression data.  Alternatively, this field may be set to ALL,
indicating that the input expression dataset is to be projected to all gene
sets defined in the specified gene set database(s). (default: \code{ALL})}

\item{min_sz}{minimal size of each geneSet for analyzing}

\item{max_sz}{maximal size of genes annotated for testing}

\item{sample_norm_type}{Normalization method applied to expression data.
Supported methods are rank, log.rank, and log.  (Default: rank)}

\item{weight}{Exponential weight employed in calculation of enrichment
scores.  The default value of \code{0.75} was selected after extensive
testing.  The module authors strongly recommend against changing from
default. (Default: \code{0.75})}

\item{min_overlap}{min overlap required between genes in gene set and genes
in input (feature dataset) file in order to include that gene set in data
set projection}

\item{BPPARAM}{see \code{\link[BiocParallel:register]{bpparam}}. Default:
\code{NULL} means \code{\cr switch (Sys.info()[["sysname"]], \cr Linux =
  BiocParallel::MulticoreParam(), \cr Windows = BiocParallel::SnowParam())}}

\item{verbose}{if TRUE, print extra infos. Default: \code{TRUE}}
}
\value{
a \code{list} or \code{ExpressionSet} or \code{SummarizedExperiment}
object of projection results for each gene set and each sample based on
\code{class(data_set)}
}
\description{
Project each sample within a data set onto a space of gene set enrichment
scores using the ssGSEA projection methodology. According to GenePattern
(modified) and GSVA
}
\references{
\itemize{\item Subramanian A, Tamayo P, Mootha VK, Mukherjee S,
Ebert BL, Gillette MA, Paulovich A, Pomeroy SL, Golub TR, Lander ES,
Mesirov JP. Gene set enrichment analysis: A knowledge-based approach for
interpreting genome-wide expression profiles. PNAS.
2005;102(43):15545-15550. \url{https://doi.org/10.1073/pnas.0506580102}
\item Barbie, D., Tamayo, P., Boehm, J. et al. Systematic RNA interference
reveals that oncogenic KRAS-driven cancers require TBK1. Nature 462,
108–112 (2009). \url{https://doi.org/10.1038/nature08460}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
