% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{stat_between_test}
\alias{stat_between_test}
\title{Between group test}
\usage{
stat_between_test(
  data = NULL,
  x,
  y,
  type = c("nonparametric", "parametric"),
  ...
)
}
\arguments{
\item{data}{a data.frame object containing the variables in the x and y or
NULL.}

\item{x, y}{if \code{data} supplied, \code{x} and \code{y} are quoted
arguments, can be a list, if \code{x} or \code{y} is a list object, each
element in \code{x} and in \code{y} will be paired to implement test; if
\code{data} is NULL, \code{x} and \code{y} should be a vector or a list
with atmomic vector in the same length.}

\item{type}{if \code{type = "nonparametric"},
\code{\link[stats]{wilcox.test}} or \code{\link[stats]{kruskal.test}} is
used. \cr if \code{type = "parametric"}, \code{\link[stats]{t.test}} or
\code{\link[stats]{aov}} is used.\cr which method is used depends on the
unique values of \code{x} in \code{data}}

\item{...}{other arguments passed to statistical test function,
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{kruskal.test}},
\code{\link[stats]{t.test}} and \code{\link[stats]{aov}}}
}
\value{
a tibble of the test results where each row match each pair of x and
y. (see \code{\link[generics:tidy]{tidy}})
}
\description{
Function to calculate statistics and P-value for both parametric and
nonparametric tests.
}
\details{
if \code{type = "nonparametric"}, \code{\link[stats]{wilcox.test}}
or \code{\link[stats]{kruskal.test}} is used. \cr if \code{type =
  "parametric"}, \code{\link[stats]{t.test}} or \code{\link[stats]{aov}} is
used.\cr The final choice of which method is used depends on the unique
values of \code{x} term.
}
\section{Quasiquotation}{
 if \code{data} supplied, \code{x} and \code{y} in
\code{stat_between_test()} are \code{\link[rlang:nse-defuse]{quoted
  arguments}} This means that its inputs are quoted to be evaluated in the
context of the data. This makes it easy to work with variables from the
data frame because you can name those directly. The flip side is that you
have to use \code{\link[rlang:nse-force]{quasiquotation}} to program with
\code{stat_between_test()}. See a tidy evaluation tutorial such as the
\href{https://dplyr.tidyverse.org/articles/programming.html}{dplyr
programming vignette} to learn more about these techniques.
}

\examples{
stat_between_test(mtcars, factor(vs), mpg, type = "p")
stat_between_test(mtcars, factor(vs), mpg, type = "n")
stat_between_test(mtcars, factor(cyl), mpg, type = "p")
stat_between_test(mtcars, factor(cyl), mpg, type = "n")
stat_between_test(
    x = lapply(mtcars[c(2, 8)], as.factor),
    y = mtcars[c(1, 3, 4)], type = "n"
)
stat_between_test(
    x = factor(mtcars[["vs"]]),
    y = mtcars["mpg"], type = "n"
)
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
