% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils_fns.R
\name{stat_cox_test}
\alias{stat_cox_test}
\title{Proportional Hazards Regression Model analysis}
\usage{
stat_cox_test(data, formula, exponentiate = TRUE, conf_level = 0.95, ...)
}
\arguments{
\item{data}{a \code{data.frame} in which to interpret the variables named in
the formula, or in the subset and the weights argument.}

\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right. The response must be a survival
object as returned by the Surv function.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the the
coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults:
\code{TRUE}. Details see \code{\link[broom:tidy.coxph]{tidy}}}

\item{conf_level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval. Details see
\code{\link[broom:tidy.coxph]{tidy}}. Default: \code{0.95}}

\item{...}{Additional arguments passed to \code{\link[survival:coxph]{coxph}}}
}
\value{
a list with two item. One for overall score test results and one for
summarized information corresponding to each term in the RHS of formula.
Details see \code{\link[broom:tidy.coxph]{tidy}}
}
\description{
Fits a Cox proportional hazards regression model and extract log-rank test
results and tidy informations. log-rank test is just a special case of the
Cox model \code{(score test)}
}
\examples{
  stat_cox_test(survival_example_data,
               survival::Surv(time, status) ~ ph.ecog + tt(age)
               )
}
\references{
\itemize{\item
\href{https://stats.stackexchange.com/questions/362381/logrank-p-value-for-2-groups}{logrank-p-value-for-2-groups}
\cr \item
\href{https://stats.stackexchange.com/questions/486806/the-logrank-test-statistic-is-equivalent-to-the-score-of-a-cox-regression-is-th}{the-logrank-test-statistic-is-equivalent-to-the-score-of-a-cox-regression-is-th}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
