% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_preprocessing_fns.R
\name{tcga_get_cli_indexed}
\alias{tcga_get_cli_indexed}
\title{download TCGA clinical data from indexed data}
\usage{
tcga_get_cli_indexed(project)
}
\arguments{
\item{project}{TCGA project ID (see \code{TCGAbiolinks::getGDCprojects()})}
}
\description{
download TCGA clinical data from indexed data
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
