% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{tcga_get_cli_xml}
\alias{tcga_get_cli_xml}
\title{download TCGA clinical data from xml files ------------------------------}
\usage{
tcga_get_cli_xml(project, path = "rawdata/GDCdata")
}
\arguments{
\item{project}{TCGA project ID (see \code{TCGAbiolinks::getGDCprojects()})}

\item{path}{Directory/Folder where the data was downloaded. Default:
\code{"rawdata/GDCdata"}}
}
\description{
download TCGA clinical data from xml files ------------------------------
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
