% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{tcga_remove_duplicated_samples}
\alias{tcga_remove_duplicated_samples}
\title{remove duplicated samples in TCGA}
\usage{
tcga_remove_duplicated_samples(barcode)
}
\arguments{
\item{barcode}{a character vector gives barcode of TCGA}
}
\value{
a tibble with duplicated barcode removed
}
\description{
remove duplicated samples in TCGA based on Firehose principle
}
\details{
In many instances there is more than one aliquot for a given
combination of individual, platform, and data type. However, only one
aliquot may be ingested into Firehose. Therefore, a set of precedence rules
are applied to select the most scientifically advantageous one among them.

The following precedence rules are applied when the aliquots have differing
analytes. For RNA aliquots, T analytes are dropped in preference to H and R
analytes, since T is the inferior extraction protocol. If H and R are
encountered, H is the chosen analyte. This is somewhat arbitrary and
subject to change, since it is not clear at present whether H or R is the
better protocol. If there are multiple aliquots associated with the chosen
RNA analyte, the aliquot with the later plate number is chosen. For DNA
aliquots, D analytes (native DNA) are preferred over G, W, or X
(whole-genome amplified) analytes, unless the G, W, or X analyte sample has
a higher plate number.
}
\references{
\href{http://gdac.broadinstitute.org/runs/stddata__2014_02_15/samples_report/LAML_Replicate_Samples.html}{Replicate Samples - GDAC Firehose}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
