% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils_fns.R
\name{valid_npc}
\alias{valid_npc}
\title{Compute npc coordinates}
\usage{
valid_npc(x)
}
\arguments{
\item{x}{numeric or character vector of coordinates.}
}
\value{
A numeric vector with values representing npc coordinates.
}
\description{
Convert character-encoded positions to npc units and keep numeric as npc
values.
}
