% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yjkit.R
\docType{package}
\name{yjkit}
\alias{yjkit}
\title{yjkit: A toolkit package for Yun and Jin}
\description{
A miscellaneous toolkit including analysis algorithm,
visualization tools and other utility functions
}
\section{Analysis Functions}{

\code{\link{run_absolute}}: Automate ABSOLUTE calling for multiple samples in
parallel ways
\code{\link{run_cibersort}}: Robust enumeration of cell subsets
from tissue expression profiles
\code{\link{run_arm_cnv}}: Caculate Chromosome-arm-levels copy number variation
\code{\link{run_ssgsea}}: ssGSEA according to GenePattern (modified) and GSVA
}

\section{Visualization Functions}{

}

