% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yjtools-package.R
\docType{package}
\name{yjtools-package}
\alias{yjtools}
\alias{yjtools-package}
\title{yjtools: Toolkit for Yun and Jin (yjtools),}
\description{
A miscellaneous toolkit including analysis algorithm, visualization tools and other utility functions. This package is just for personal use.
}
\keyword{internal}
