% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yjtools-package.R
\docType{package}
\name{yjtools}
\alias{yjtools}
\alias{yjtools-package}
\title{yjtools: A toolkit package for Yun and Jin}
\description{
A miscellaneous toolkit including analytical algorithm,
visualization tools and other utility functions.
}
\section{Visualization Functions}{
 \code{\link{ggannotate_npc}}: Create an
ggplot2 annotation layer.\cr \cr \code{\link{geom_label_npc}} and
\code{\link{geom_text_npc}}: ggplot2 layer for Text or label put in
Normalised Parent Coordinates. \cr
}

\section{Statistical Functions}{
 \code{\link{stat_between_test}}: Between
group test.\cr \cr \code{\link{stat_cor_test}}: Correlation analysis.\cr
}

\author{
Yun \email{yunyunpp96@outlook.com}
}
\keyword{internal}
