% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstat_utils_fns.R
\name{gganno_between_test}
\alias{gganno_between_test}
\title{Create a ggplot2 object with annotation for group or condition test results}
\usage{
gganno_between_test(
  data,
  mapping = aes(x, y, ...),
  type = c("nonparametric", "parametric"),
  format_label = TRUE,
  label_position = c(0.02, 1),
  label_justification = NULL,
  anno_statistic = TRUE,
  label_sep = NULL,
  ggtheme = ggthemes::theme_few(),
  anno_args = list(fontface = "bold"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not specified, must be
supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. x and y
aesthetic must in.}

\item{type}{if \code{type = "nonparametric"},
\code{\link[stats]{wilcox.test}} or \code{\link[stats]{kruskal.test}} is
used. \cr if \code{type = "parametric"}, \code{\link[stats]{t.test}} or
\code{\link[stats]{aov}} is used.\cr which method is used depends on the
unique values of \code{x} in \code{data}}

\item{format_label}{if \code{TRUE}, will be formatted by
\code{\link{format_num}}. Default: \code{TRUE}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length 2 or a scalar character vector \code{(one of
  "title", "subtitle", "caption", and "tag")}.}

\item{label_justification}{These can either be a number between 0
(\code{right/bottom}) and 1 (\code{top/left}) or a character
("\code{left}", "\code{middle}", "\code{right}", "\code{bottom}",
"\code{center}", "\code{top}"). There are two special alignments:
"\code{inward}" and "\code{outward}". Inward always aligns text towards the
center, and outward aligns it away from the center.}

\item{anno_statistic}{a scalar logical, Should test statistic be annotated in
the label}

\item{label_sep}{see argument \code{label}. String to insert between the test
statistic and test P-value.}

\item{ggtheme}{a ggplot2 theme object. Provided as a ggplot2 theme object or
a function (can be a function name as a string) implemented as a ggplot2
theme object.}

\item{anno_args}{additional arguments control the \code{label}. See
\code{\link[grid]{textGrob}} and \code{\link[ggplot2]{element_text}}}

\item{...}{other arguments passed to statistical test function,
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{kruskal.test}},
\code{\link[stats]{t.test}} and \code{\link[stats]{aov}}}
}
\value{
a ggplot2 object with annotation of group or condition comparisons
test results
}
\description{
annotate group or condition comparisons in between-subjects designs in a
ggplot2 object.
}
\examples{
gganno_between_test(mtcars, aes(factor(cyl), mpg)) +
  ggplot2::geom_boxplot()
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
