% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstat_utils_fns.R
\name{gganno_cor_test}
\alias{gganno_cor_test}
\title{annotate Correlation coefficient in a ggplot2 object}
\usage{
gganno_cor_test(
  data,
  mapping = NULL,
  use = c("pairwise.complete.obs", "everything", "all.obs", "complete.obs",
    "na.or.complete"),
  method = c("spearman", "pearson", "kendall"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  conf_level = 0.95,
  format_label = TRUE,
  label_position = c(0.02, 0.98),
  label_justification = c(0, 1),
  anno_statistic = TRUE,
  label_sep = NULL,
  anno_args = list(fontface = "bold"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not specified, must be
supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. x and y
aesthetic must in.}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an
abbreviation of) one of the strings \code{"everything"}, \code{"all.obs"},
\code{"complete.obs"}, \code{"na.or.complete"}, or
\code{"pairwise.complete.obs"}. Default: \code{"pairwise.complete.obs"}.
See \code{\link[stats]{cor}}}

\item{method}{a character string indicating which correlation coefficient (or
covariance) is to be computed. One of \code{"pearson"}, \code{"kendall"},
or \code{"spearman"}, can be abbreviated. Default: \code{"spearman"}}

\item{alternative}{indicates the alternative hypothesis and must be one of
\code{"two.sided"}, \code{"greater"} or \code{"less"}. You can specify just
the initial letter. \code{"greater"} corresponds to positive association,
\code{"less"} to negative association.}

\item{exact}{a logical indicating whether an exact p-value should be
computed. Used for Kendall's tau and Spearman's rho. See
\code{\link[stats]{cor.test}}. Default: \code{TRUE}.}

\item{conf_level}{confidence level for the returned confidence interval.
Currently only used for the Pearson product moment correlation coefficient
if there are at least 4 complete pairs of observations.}

\item{format_label}{if \code{TRUE}, numeric label will be formatted by
\code{\link{format_num}}. Default: \code{TRUE}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length two or a character vector with length one (one of
\code{"lefttop", "righttop", "leftbottom", and "rightbottom"}) or length
two (See below \code{label_justification}). Default: \code{c(0.02, 0.98)}.}

\item{label_justification}{label_justification must be a element-two (or one)
character or numeric vector. Possible string values are: "left", "right",
"centre", "center", "bottom", and "top". For numeric values, 0 means left
(bottom) alignment and 1 means right (top) alignment. Default: \code{c(0,
  1)}.}

\item{anno_statistic}{a scalar logical, Should test statistic added in label}

\item{label_sep}{see argument \code{label}. String to insert between the test
statistic and test P-value.}

\item{anno_args}{Other arguments passed on to \code{\link[ggplot2]{layer}}.
These are often aesthetics, used to set an aesthetic to a fixed value, like
colour = "red" or size = 3. They may also be parameters to the paired
geom/stat. See \code{\link[ggplot2:geom_text]{geom_text}}}

\item{...}{Other parameters for \code{\link[stats]{cor.test}}}
}
\description{
Create a ggplot2 object with annotation for Correlation coefficient
}
\examples{
  gganno_cor_test(mtcars, aes(disp, mpg)) +
    ggplot2::geom_point()+
    ggplot2::geom_smooth()
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
