% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate_npc.R
\name{ggannotate_npc}
\alias{ggannotate_npc}
\title{Create an annotation layer with text or label annotation}
\usage{
ggannotate_npc(
  geom = "text",
  label,
  label_position = "lefttop",
  label_justification = NULL,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{geom}{character Name of geom to use for annotation. Can be one of
\code{"text"} and \code{"label"}.}

\item{label}{the character which will be added to the plot, will be connected
by \code{stringr::str_c(names(label), label), collapse = label_sep, sep =
  ": "}}

\item{label_position}{the position where the label will added, can be a
numeric vector of length two or a character vector with length one (one of
\code{"lefttop", "righttop", "leftbottom", and "rightbottom"}) or length
two (See below \code{label_justification}). Default: \code{"lefttop"}.}

\item{label_justification}{label_justification must be a element-two (or one)
character or numeric vector. Possible string values are: "left", "right",
"centre", "center", "bottom", and "top". For numeric values, 0 means left
(bottom) alignment and 1 means right (top) alignment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}}. These
are often aesthetics, used to set an aesthetic to a fixed value, like
colour = "red" or size = 3. They may also be parameters to the paired
geom/stat. See \code{\link[ggplot2:geom_text]{geom_text}} and
\code{\link[ggplot2:geom_text]{geom_label}}.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}
}
\value{
a ggplot2 layer object with annotation
}
\description{
This function Creates an ggplot2 object with annotation.
}
\examples{
ggannotate_npc(label = "a", label_position = c(0.5, 0.5))
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
