% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils_fns.R
\name{ggannotate_npc}
\alias{ggannotate_npc}
\title{Create an annotation layer}
\usage{
ggannotate_npc(geom, x = NULL, y = NULL, ..., na.rm = FALSE)
}
\arguments{
\item{geom}{character Name of geom to use for annotation. Can be one of
\code{"grob"}, \code{"plot"}, \code{"text"}, \code{"label"}, and
\code{"table"}. Details see \code{\link[ggpmisc]{annotate}}}

\item{x, y}{the position of added label in Normalised Parent
Coordinates(\code{npc})}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}}.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}
}
\description{
This function adds geoms to a plot with x and y mapping to Normalised Parent
Coordinates
}
\details{
See \code{\link[ggplot2]{annotate}}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
