% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Install or update CRAN, Bioconductor and GitHub packages}
\usage{
install_pkgs(
  pkgs = character(),
  ...,
  type = "source",
  INSTALL_opts = "--byte-compile",
  site_repository = character(),
  update = TRUE,
  ask = TRUE,
  checkBuilt = FALSE
)
}
\arguments{
\item{pkgs}{\code{character()} vector of package names to install or
update.  A missing value updates installed packages according
to \verb{update =} and \verb{ask =}. Package names containing a '/' are
treated as GitHub repositories and installed using
\code{remotes::install_github()}.}

\item{...}{Additional arguments used by \code{install.packages()}.}

\item{site_repository}{(Optional) \code{character(1)} vector
representing an additional repository in which to look for
packages to install. This repository will be prepended to the
default repositories (which you can see with
\verb{BiocManager::\\link{repositories}()}).}

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{ask}{\code{logical(1)} indicating whether to prompt user before
installed packages are updated.  If TRUE, user can choose
whether to update all outdated packages without further
prompting, to pick packages to update, or to cancel updating
(in a non-interactive session, no packages will be updated
unless \code{ask = FALSE}).}

\item{checkBuilt}{\code{logical(1)}. If \code{TRUE} a package built under an
earlier major.minor version of R (e.g., 3.4) is considered to
be old.}
}
\value{
\code{BiocManager::install()} returns the \code{pkgs} argument, invisibly.
}
\description{
Install or update CRAN, Bioconductor and GitHub packages from source and
compiling packages
}
\details{
Installation of \emph{Bioconductor} and CRAN packages use R's standard
functions for library management -- \code{install.packages()},
\code{available.packages()}, \code{update.packages()}. Installation of GitHub
packages uses the \code{remotes::install_github()}.

When installing CRAN or \emph{Bioconductor} packages, typical arguments
include: \code{lib.loc}, passed to \verb{\\link{old.packages}()} and used to
determine the library location of installed packages to be updated;
and \code{lib}, passed to \verb{\\link{install.packages}()} to determine the
library location where \code{pkgs} are to be installed.

When installing GitHub packages, \code{...} is passed to the
\pkg{remotes} package functions \verb{\\link[remotes]{install_github}()}
and \code{remotes:::install()}. A typical use is to build vignettes, via
\verb{dependencies=TRUE, build_vignettes=TRUE}.

See \code{?repositories} for additional detail on customizing where
BiocManager searches for package installation.

\env{BIOCONDUCTOR_ONLINE_VERSION_DIAGNOSIS} is an environment
variable or global \code{options()} which, when set to \code{FALSE}, avoids
the R and \emph{Bioconductor} version checks that are done by querying
an online configuration file. Setting
\env{BIOCONDUCTOR_ONLINE_VERSION_DIAGNOSIS} to \code{FALSE} can speed
package loading when internet access is slow or non-existent, but
may result in out-of-date information about the current release and
development versions of \emph{Bioconductor}.
}
\seealso{
\verb{BiocManager::\\link{repositories}()} returns the \emph{Bioconductor} and
CRAN repositories used by \code{install()}.

\verb{\\link{install.packages}()} installs the packages themselves (used by
\code{BiocManager::install} internally).

\verb{\\link{update.packages}()} updates all installed packages (used by
\code{BiocManager::install} internally).

\verb{\\link{chooseBioCmirror}()} allows choice of a mirror from all
public \emph{Bioconductor} mirrors.

\verb{\\link{chooseCRANmirror}()} allows choice of a mirror from all
public CRAN mirrors.
}
