% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils_fns.R
\name{stat_between_test}
\alias{stat_between_test}
\title{Between group test}
\usage{
stat_between_test(data, x, y, type = c("nonparametric", "parametric"), ...)
}
\arguments{
\item{data}{a data.frame object containing the variables in the x and y}

\item{x, y}{quoted arguments, can be a list, if x or y is a list object, each
of x and each of y will be paired to implement test}

\item{type}{if \code{type = "nonparametric"},
\code{\link[stats]{wilcox.test}} or \code{\link[stats]{kruskal.test}} is
used. \cr if \code{type = "parametric"}, \code{\link[stats]{t.test}} or
\code{\link[stats]{aov}} is used.\cr which method is used depends on the
unique values of \code{x} in \code{data}}

\item{...}{other arguments passed to statistical test function,
\code{\link[stats]{wilcox.test}}, \code{\link[stats]{kruskal.test}},
\code{\link[stats]{t.test}} and \code{\link[stats]{aov}}}
}
\value{
a tibble of the test results where each row match each pair of x and
y. (see \code{\link[generics:tidy]{tidy}})
}
\description{
Function to calculate statistics and P-value for both parametric and
nonparametric tests.
}
\details{
if \code{type = "nonparametric"}, \code{\link[stats]{wilcox.test}}
or \code{\link[stats]{kruskal.test}} is used. \cr if \code{type =
  "parametric"}, \code{\link[stats]{t.test}} or \code{\link[stats]{aov}} is
used.\cr which method is used depends on the unique values of \code{x term}
in \code{data}
}
\section{Quasiquotation}{

\code{x} and \code{y} in \code{stat_between_test()}
are \code{\link[rlang:nse-defuse]{quoted arguments}} This means that its
inputs are quoted to be evaluated in the context of the data. This makes it
easy to work with variables from the data frame because you can name those
directly. The flip side is that you have to use
\code{\link[rlang:nse-force]{quasiquotation}} to program with
\code{stat_between_test()}. See a tidy evaluation tutorial such as the
\href{https://dplyr.tidyverse.org/articles/programming.html}{dplyr
programming vignette} to learn more about these techniques.
}

\examples{
  stat_between_test(mtcars, vs, mpg, type = "p")
  stat_between_test(mtcars, vs, mpg, type = "n")
  stat_between_test(mtcars, cyl, mpg, type = "p")
  stat_between_test(mtcars, cyl, mpg, type = "n")
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
