% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils_fns.R
\name{stat_cor_test}
\alias{stat_cor_test}
\title{Correlation analysis}
\usage{
stat_cor_test(
  x,
  y = NULL,
  use = c("pairwise.complete.obs", "everything", "all.obs", "complete.obs",
    "na.or.complete"),
  method = c("spearman", "pearson", "kendall"),
  alternative = c("two.sided", "less", "greater"),
  exact = TRUE,
  cor_test = FALSE,
  ...
)
}
\arguments{
\item{x, y}{data.frame object. x and y must have the same \code{nrow}}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an
abbreviation of) one of the strings \code{"everything"}, \code{"all.obs"},
\code{"complete.obs"}, \code{"na.or.complete"}, or
\code{"pairwise.complete.obs"}. Default: \code{"pairwise.complete.obs"}.
See \code{\link[stats]{cor}}.}

\item{method}{a character string indicating which correlation coefficient (or
covariance) is to be computed. One of \code{"pearson"}, \code{"kendall"},
or \code{"spearman"}, can be abbreviated. Default: \code{"spearman"}}

\item{alternative}{indicates the alternative hypothesis and must be one of
\code{"two.sided"}, \code{"greater"} or \code{"less"}. You can specify just
the initial letter. \code{"greater"} corresponds to positive association,
\code{"less"} to negative association.}

\item{exact}{a logical indicating whether an exact p-value should be
computed. Used for Kendall's tau and Spearman's rho. See
\code{\link[stats]{cor.test}}. Default: \code{TRUE}.}

\item{cor_test}{a logical value indicates whether the estimation of P-value
should use \code{\link[stats]{cor.test}} instead. If FALSE, the estimation
of P-value will use \code{\link[stats:TDist]{stats::pt()}} (for spearman or pearson) or
\code{\link[stats:Normal]{stats::pnorm()}} (for kendall). Default: \code{FALSE}.}

\item{...}{Other parameters passed to \code{\link[stats]{cor.test}}.}
}
\value{
a tibble of the Correlation results. (see
\code{\link[stats]{cor.test}}).
}
\description{
Function to calculate Correlation coefficient and P-value between the columns
of x and the columns of y, using one of Pearson's product moment correlation
coefficient, Kendall's tau or Spearman's rho.
}
\details{
See \code{\link[stats]{cor.test}} and \code{\link[stats]{cor}}.
}
\examples{
  stat_cor_test(mtcars)
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
